Pool VT on a hotpath
